enable hud for non-gimbling cameras
